[백단 및 프론트엔드 관련 변경 사항]
<백단 변경 사항>
필터를 통한 데이터 송수신 xss처리
회원가입 기능 추가
식당 후기 작성 기능 추가

<프론트단 변경 사항>
회원가입 화면 추가
드롭메뉴 추가 및 로그아웃 위치 변경

+++ src/main/java/com/main/vo/Comment.java
@@ -30,7 +30,7 @@ public class Comment {
 	@Column(name = "place_id")
 	private Integer placeId;
 	@Column(name = "user_seq")
-	private Integer userSeq;
+	private Long userSeq;
 	private String comment;
 	private double rating;
 	@Column(name = "image_path")
